Extracted FlatList component to a HOC for the list
